LayoutTests/editing/execCommand/strikethroughSelection.html

<html> 
<head>

<style>
.editing { 
    border: 2px solid red; 
    padding: 12px; 
    font-size: 24px; 
}
</style>
<script src=../editing.js language="JavaScript" type="text/JavaScript" ></script>

<script>

function editingTest() {

    for (i = 0; i < 21; i++) {
            execExtendSelectionForwardByCharacterCommand();
            execStrikethroughCommand();
            
    }
 }

</script>

<title>Editing Test</title> 
</head> 
<body>
<p>This test striking through the first 21st letters.  Every letter including spaces before "jumps" must be striked.</p>
<div contenteditable id="root" class="editing">
<span id="test">"The quick brown fox jumps over the lazy dog" uses every
letter in the english language.
</span>
</div>

<script>
runDumpAsTextEditingTest();
</script>

</body>
</html>
