LayoutTests/editing/execCommand/strikethroughSelection-expected.txt

This test striking through the first 21st letters. Every letter including spaces before "jumps" must be striked.

"The quick brown fox jumps over the lazy dog" uses every letter in the english language.
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"</span>The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"Th</span>e quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The </span>quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The qu</span>ick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The quic</span>k brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The quick </span>brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The quick br</span>own fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The quick brow</span>n fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The quick brown </span>fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The quick brown fo</span>x jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test">"The quick brown fox jumps over the lazy dog" uses every letter in the english language. </span>
execStrikethroughCommand: <span id="test"><span class="Apple-style-span" style="text-decoration: line-through;">"The quick brown fox </span>jumps over the lazy dog" uses every letter in the english language. </span>
